+++ src/Exceptions/CannotFindRouteException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Glhd\Gretel\Exceptions;
+
+use InvalidArgumentException;
+
+class CannotFindRouteException extends InvalidArgumentException
+{
+	public function __construct(string $name)
+	{
+		parent::__construct("Unable to find a route named '{$name}'.");
+	}
+}