+++ src/Registry.php
@@ -3,9 +3,14 @@
 namespace Glhd\Gretel;
 
 use Closure;
+use Glhd\Gretel\Exceptions\CannotFindRouteException;
 use Glhd\Gretel\Exceptions\MissingBreadcrumbException;
+use Glhd\Gretel\Resolvers\ParentResolver;
+use Glhd\Gretel\Resolvers\TitleResolver;
+use Glhd\Gretel\Resolvers\UrlResolver;
 use Glhd\Gretel\Routing\RouteBreadcrumb;
 use Illuminate\Routing\Route;
+use Illuminate\Routing\Router;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Traits\ForwardsCalls;
 use Throwable;
@@ -21,12 +26,15 @@ class Registry
 	
 	protected const HANDLER_MISCONFIGURED = 'misconfigured';
 	
+	protected Router $router;
+	
 	protected Collection $breadcrumbs;
 	
 	protected Collection $exception_handlers;
 	
-	public function __construct()
+	public function __construct(Router $router)
 	{
+		$this->router = $router;
 		$this->breadcrumbs = new Collection();
 		$this->exception_handlers = new Collection();
 		
@@ -94,6 +102,17 @@ class Registry
 		return $this;
 	}
 	
+	public function breadcrumb(string $name, $title = null, $parent = null, $relation = null)
+	{
+		$route = $this->getRouteByName($name);
+		
+		$title = TitleResolver::make($title);
+		$parent = ParentResolver::make($parent, $name, $relation);
+		$url = UrlResolver::make($name, $route->parameterNames());
+		
+		return $this->register(new RouteBreadcrumb($name, $title, $parent, $url));
+	}
+	
 	public function register(RouteBreadcrumb ...$breadcrumbs): Registry
 	{
 		foreach ($breadcrumbs as $breadcrumb) {
@@ -131,6 +150,25 @@ class Registry
 		return $route instanceof Route
 			? $route->getName()
 			: (string) $route;
+	}
+	
+	protected function getRouteByName(string $name, bool $refresh = false): ?Route
+	{
+		$routes = $this->router->getRoutes();
+		
+		if ($refresh) {
+			$routes->refreshNameLookups();
+		}
+		
+		if ($route = $routes->getByName($name)) {
+			return $route;
+		}
+		
+		if (! $refresh) {
+			return $this->getRouteByName($name, true);
+		}
+		
+		throw new CannotFindRouteException($name);
 	}
 	
 	/**